Split solver folder into solver, constraints, state, and step folders
