+++ src/restruct_poc/constraints/calculate_constraint_residual_gradient.hpp
