+++ src/restruct_poc/constraints/calculate_rigid_constraint.hpp
