+++ src/restruct_poc/constraints/calculate_rotation_control_constraint.hpp
