+++ src/restruct_poc/constraints/constraint.hpp
@@ -1,29 +1,12 @@
 #pragma once
 
+#include "constraint_type.hpp"
+
 #include "src/restruct_poc/math/vector_operations.hpp"
 #include "src/restruct_poc/model/node.hpp"
 
+
 namespace openturbine {
-
-/// @brief Enum class to define the type of constraint
-enum class ConstraintType : std::uint8_t {
-    kNone = 0,          // No constraint (default)
-    kFixedBC = 1,       // Fixed boundary condition constraint (zero displacement)
-    kPrescribedBC = 2,  // Prescribed boundary condition (displacement can be set)
-    kRigid =
-        3,  // Rigid constraint between two nodes (nodes maintain relative distance and rotation)
-    kCylindrical = 4,      // Target node rotates freely around specified axis. Relative distance and
-                           // rotation are fixed)
-    kRotationControl = 5,  // Specify rotation about given axis
-};
-
-KOKKOS_INLINE_FUNCTION
-constexpr size_t GetNumberOfNodes(ConstraintType t) {
-    const auto has_two_nodes = t == ConstraintType::kRigid || t == ConstraintType::kCylindrical ||
-                               t == ConstraintType::kRotationControl;
-    return 1U + static_cast<size_t>(has_two_nodes);
-}
-
 /// @brief Struct to define a constraint between two nodes
 /// @details A constraint is a relationship between two nodes that restricts their relative motion
 /// in some way. Constraints can be used to model fixed boundary conditions, prescribed
