+++ src/restruct_poc/solver/populate_sparse_row_ptrs_col_inds_transpose.hpp
@@ -2,7 +2,7 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "constraints.hpp"
+#include "src/restruct_poc/constraints/constraint_type.hpp"
 
 #include "src/restruct_poc/types.hpp"
