+++ src/restruct_poc/state/calculate_next_state.hpp
