+++ src/restruct_poc/state/copy_nodes_to_state.hpp
