+++ src/restruct_poc/state/update_algorithmic_acceleration.hpp
