+++ src/restruct_poc/state/update_static_prediction.hpp
