+++ src/restruct_poc/step/assemble_constraints_matrix.hpp
@@ -3,10 +3,10 @@
 #include <Kokkos_Core.hpp>
 #include <Kokkos_Profiling_ScopedRegion.hpp>
 
-#include "solver.hpp"
-#include "constraints.hpp"
-#include "copy_constraints_to_sparse_matrix.hpp"
-#include "copy_sparse_values_to_transpose.hpp"
+#include "src/restruct_poc/solver/solver.hpp"
+#include "src/restruct_poc/constraints/constraints.hpp"
+#include "src/restruct_poc/solver/copy_constraints_to_sparse_matrix.hpp"
+#include "src/restruct_poc/solver/copy_sparse_values_to_transpose.hpp"
 
 namespace openturbine {
 inline void AssembleConstraintsMatrix(Solver& solver, Constraints& constraints) {
