+++ src/restruct_poc/step/assemble_stiffness_matrix.hpp
@@ -3,10 +3,9 @@
 #include <Kokkos_Core.hpp>
 #include <Kokkos_Profiling_ScopedRegion.hpp>
 
-#include "integrate_stiffness_matrix.hpp"
+#include "src/restruct_poc/system/integrate_stiffness_matrix.hpp"
 
 #include "src/restruct_poc/beams/beams.hpp"
-#include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
