+++ src/restruct_poc/step/assemble_tangent_operator.hpp
@@ -3,10 +3,10 @@
 #include <Kokkos_Core.hpp>
 #include <Kokkos_Profiling_ScopedRegion.hpp>
 
-#include "solver.hpp"
-#include "state.hpp"
+#include "src/restruct_poc/solver/solver.hpp"
+#include "src/restruct_poc/solver/copy_tangent_to_sparse_matrix.hpp"
+#include "src/restruct_poc/state/state.hpp"
 #include "src/restruct_poc/system/calculate_tangent_operator.hpp"
-#include "copy_tangent_to_sparse_matrix.hpp"
 
 namespace openturbine {
