+++ src/restruct_poc/step/update_state_prediction.hpp
@@ -3,12 +3,14 @@
 #include <Kokkos_Core.hpp>
 #include <Kokkos_Profiling_ScopedRegion.hpp>
 
-#include "calculate_displacement.hpp"
-#include "solver.hpp"
-#include "state.hpp"
 #include "step_parameters.hpp"
-#include "update_dynamic_prediction.hpp"
-#include "update_static_prediction.hpp"
+
+#include "src/restruct_poc/state/calculate_displacement.hpp"
+#include "src/restruct_poc/state/state.hpp"
+#include "src/restruct_poc/state/update_dynamic_prediction.hpp"
+#include "src/restruct_poc/state/update_static_prediction.hpp"
+
+#include "src/restruct_poc/solver/solver.hpp"
 
 #include "src/restruct_poc/types.hpp"
